% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_heckman_binary.R
\name{loglik_bivprob_est}
\alias{loglik_bivprob_est}
\alias{loglik_bivprob}
\alias{loglik_bivprob_grad}
\title{Internal function}
\usage{
loglik_bivprob_est(data, yS, yO, XO, XS, kS, kO, sel, out, ...)

loglik_bivprob(param, yS, yO, XO, XS, kS, kO)

loglik_bivprob_grad(param, yS, yO, XO, XS, kS, kO)
}
\description{
This function is a wrapper for estimating the censored bivariate probit model. Calculates starting values and conducts numerical optimization. The other two functions calculate the log likelihood and analytic gradients.
}
\keyword{internal}
