% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivprob_lik_function_ghq.R
\name{loglik_bivprob_ghq}
\alias{loglik_bivprob_ghq}
\title{Internal function}
\usage{
loglik_bivprob_ghq(param, yS, yO, XO, XS, kS, kO, group, QP, ...)
}
\description{
This function calculates the log likelihood of the censored bivariate probit model with random intercept using GHQ. Calls a C++ function to speed up the loop.
}
\keyword{internal}
