% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordbivprob_est_function_ghq.R
\name{loglik_ordbivprob_est_ghq}
\alias{loglik_ordbivprob_est_ghq}
\title{Internal function}
\usage{
loglik_ordbivprob_est_ghq(
  data,
  yS,
  yO,
  XO,
  XS,
  kS,
  kO,
  group,
  sel,
  out,
  QP,
  ...
)
}
\description{
This function is a wrapper for estimating the ordered probit model with sample selection and random intercept using GHQ. Calculates starting values and conducts numerical optimization.
}
\keyword{internal}
