% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordbivprob_lik_function_ghq.R
\name{loglik_ordbivprob_ghq}
\alias{loglik_ordbivprob_ghq}
\title{Internal function}
\usage{
loglik_ordbivprob_ghq(param, yS, yO, XO, XS, kS, kO, group, QP, ...)
}
\description{
This function calculates the log likelihood of the ordered probit model with sample selection and random intercept using GHQ. Calls a C++ function to speed up the loop.
}
\keyword{internal}
