% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mode}
\alias{mode}
\alias{loop1}
\title{Internal functions written in C++ to speed up maximization process of the censored bivariate probit model with random intercepts}
\usage{
mode(par, x, yS, yO, linpredS, linpredO, group_aux, rho, tau, sigma1, sigma2)

loop1(
  yS,
  yO,
  linpredS,
  linpredO,
  rho,
  tau,
  sigma1,
  sigma2,
  w1,
  a1,
  group_aux,
  group,
  M,
  chol_vc_re
)
}
\description{
Functions that calculate the modes of the random intercepts and that contain the loops for computing the log likelihood approximated by GHQ.
}
\keyword{internal}
