% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mode_ord}
\alias{mode_ord}
\alias{loop1_ord}
\title{Internal functions written in C++ to speed up maximization process of the ordered probit model with sample selection and random intercepts}
\usage{
mode_ord(
  par,
  x,
  yS,
  yO,
  linpredS,
  linpredO,
  group_aux,
  rho,
  tau,
  z,
  sigma1,
  sigma2,
  levels
)

loop1_ord(
  yS,
  yO,
  linpredS,
  linpredO,
  rho,
  tau,
  sigma1,
  sigma2,
  z,
  levels,
  w1,
  a1,
  group_aux,
  group,
  M,
  chol_vc_re
)
}
\description{
Functions that calculate the modes of the random intercepts and that contain the loops for computing the log likelihood approximated by GHQ.
}
\keyword{internal}
